requirements.txt
# Core dependencies
google-generativeai>=0.3.0
PyPDF2>=3.0.1
python-docx>=0.8.11

# Async support
aiohttp>=3.8.5

# Utility libraries
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: For advanced embeddings and vector search
# sentence-transformers>=2.2.2
# chromadb>=0.4.22
# faiss-cpu>=1.7.4

# For the Streamlit dashboard (optional)
# streamlit>=1.29.0
# plotly>=5.18.0